auth-proxy/requirements.txt
# Auth Proxy Requirements

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP client
httpx==0.26.0

# Google authentication
google-auth==2.27.0
